model-first baseline
